src/room/api/serializers.py
from rest_framework import serializers

from room.models import Room


class RoomSerializer(serializers.ModelSerializer):
    created_by = serializers.ReadOnlyField()
    updated_by = serializers.ReadOnlyField()
    created_at = serializers.ReadOnlyField()
    updated_at = serializers.ReadOnlyField()

    class Meta:
        model = Room
        fields = '__all__'
